% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bangdiwala_chart.R
\name{bangdiwala_chart}
\alias{bangdiwala_chart}
\title{Bangdiwala Chart}
\usage{
bangdiwala_chart(x, main = "", xlab = names(dimnames(x))[2],
  ylab = names(dimnames(x))[1], cat.label = colnames(x), las = TRUE,
  add.legend = TRUE, location.legend = "bottomright",
  add.B.hat = TRUE, location.B.hat = "topleft", add.x.margin = TRUE,
  add.y.margin = TRUE, ordinal.data = TRUE, col.line = "grey",
  lwd.line = 2, lty.line = 1)
}
\arguments{
\item{x}{data as a table object in R.
note: columns of x will turn into x-axis,
rows of x will turn into y-axis,
to reverse, use \code{t()} function (i.e., transpose).}

\item{main}{title of plot.}

\item{xlab}{title of x-axis (default: extracted from x).}

\item{ylab}{title of y-axis (default: extracted from x).}

\item{cat.label}{category labels (default: extracted from x).}

\item{las}{y-axis category labels rotated.}

\item{add.legend}{`TRUE/FALSE`, add a legend to the plot.}

\item{location.legend}{either `bottomright` or `topleft` (only relevant if `add.legend=TRUE`).}

\item{add.B.hat}{`TRUE/FALSE`, add Bangdiwala agreement measure to plot,
(otherwise it is only given as the function's output),
(agreement measure without weights).}

\item{location.B.hat}{either `bottomright` or `topleft` (only relevant if `add.B.hat=TRUE`).}

\item{add.x.margin}{`TRUE/FALSE`, add marginal values for columns (x) to plot.}

\item{add.y.margin}{`TRUE/FALSE`, add marginal values for columns (y) to plot.}

\item{ordinal.data}{`TRUE/FALSE`, if `TRUE`, partial agreement squares are added.}

\item{col.line}{color of line (default: `firebrick`).}

\item{lwd.line}{width of line (default: `2`).}

\item{lty.line}{type of line (`solid`, `dotted`, etc.; default: `solid`).}
}
\value{
A chart, and Bangdiwala agreement measure
}
\description{
A function for creating a Bangdiwala Chart.
   Constructed using "The agreement chart" by Shrikant I Bangdiwala.
   To learn more about Bangdiwala Chart, see
   \url{http://www.stat.ncsu.edu/information/library/mimeo.archive/ISMS_1988_1859.pdf}, and
   book: \url{http://www.utdallas.edu/~pkc022000/agreement_book/} chapter 12: Categorical Data.
}
\references{
Bangdiwala, S. I. (1988). The Agreement Chart. Department of Biostatistics, University of North Carolina at Chapel Hill, Institute of Statistics Mimeo Series No. 1859, http://www.stat.ncsu.edu/information/library/mimeo.archive/ISMS_1988_1859.pdf

     David Meyer, Achim Zeileis, and Kurt Hornik (2017). vcd: Visualizing Categorical Data. R package version 1.4-4.

     Choudhary, P. K. and Nagaraja, H. N. (2017). Measuring Agreement: Models, Methods, and Applications, Wiley.

     Westlund, K. B. and Kurland, L. T. (1953). Studies on multiple sclerosis in Winnipeg, Manitoba, and New Orleans, Louisiana I. Prevalence; comparison between the patient groups in Winnipeg and New Orleans. American Journal of Hygiene 57, 380–396.
}
